+++ src/app/locales/en/components/qos-modal/file-entry.js
@@ -8,6 +8,7 @@ const translations = {
   statusHint: {
     fulfilled: 'All requirements for this {{fileType}} are fulfilled',
     pending: 'Pending ‐ there are some unfulfilled requirements for this {{fileType}}',
+    impossible: 'At least one requirement is impossible to be fulfilled',
     empty: 'No QoS requirements defined for this {{fileType}}',
     error: 'An error occured when computing this {{fileType}}\'s QoS status ‐ expand QoS requirements panel for details',
   },