+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
 ### Latest changes
 
+* VFS-6344 Added support for impossible QoS requirement and lack of QoS privileges
 * VFS-6358 Uploader optimization
 * VFS-6381 Fixed build process
 * VFS-6343 Updated common libs